add Encoder wrapper, fewer pointers
